vp: core: mmu: fixed translation for >32 bit addresses
The physical page number (PPN) was calculated by a 32 instead of a 64
bit mask.

Problem:
 * Original PPN calculation: PPN = (paddr & ~PGMASK);
 * PGMASK is of type "unsigned", which is 32 bit:
   Value: 0b1111_1111_1111 (for 4k pages - 12 bits)
 * For calculating a 64 bit PPN, we need a *64 bit* mask
   where all bits except the last 12 are set
   (0b1111_1111_1111_1111_1111_1111_1111_1111__1111_1111_1111_1111_1111_0000_0000_0000)
   To get this mask, PGMASK is inverted bitwise -> ~PGMASK
 * However, the inverted PGMASK is still *32 bit*
   (0b1111_1111_1111_1111_1111_0000_0000_0000)
   Bits 32-64 are "missing" and treated as 0.
   -> Error in PPN calcuation.

Fix:
Cast the mask to 64 bit before inverting it -> 64 bit mask.

+++ vp/src/core/common/mmu.h
@@ -121,7 +121,7 @@ struct MMU_T {
 		uint64_t paddr = walk(vaddr, type, mode);
 
 		// optimization only, to void page walk
-		x.ppn = (paddr & ~PGMASK);
+		x.ppn = (paddr & ~((uint64_t)PGMASK));
 		x.vpn = vpn;
 
 		return paddr;
